refactor(e2e): refine e2e test about timechaos
