+++ test/e2e/chaos/timechaos/misc.go
@@ -1,0 +1,28 @@
+package timechaos
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"time"
+)
+
+// get pod current time in nanosecond
+func getPodTimeNS(c http.Client, port uint16) (*time.Time, error) {
+	resp, err := c.Get(fmt.Sprintf("http://localhost:%d/time", port))
+	if err != nil {
+		return nil, err
+	}
+
+	out, err := ioutil.ReadAll(resp.Body)
+	defer resp.Body.Close()
+	if err != nil {
+		return nil, err
+	}
+
+	t, err := time.Parse(time.RFC3339Nano, string(out))
+	if err != nil {
+		return nil, err
+	}
+	return &t, nil
+}
