+++ test/e2e/chaos/timechaos/time_skew.go
@@ -1,0 +1,204 @@
+package timechaos
+
+import (
+	"context"
+	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
+	"github.com/chaos-mesh/chaos-mesh/test/e2e/util"
+	. "github.com/onsi/ginkgo"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/kubernetes/test/e2e/framework"
+	"k8s.io/utils/pointer"
+	"net/http"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"time"
+)
+
+func TestcaseTimeSkewOnceThenRecover(
+	ns string,
+	cli client.Client,
+	c http.Client,
+	port uint16,
+) {
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	By("wait e2e helper ready")
+	err := util.WaitE2EHelperReady(c, port)
+	framework.ExpectNoError(err, "wait e2e helper ready error")
+
+	By("create chaos CRD objects")
+	initTime, err := getPodTimeNS(c, port)
+	framework.ExpectNoError(err, "failed to get pod time")
+
+	timeChaos := &v1alpha1.TimeChaos{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "timer-time-chaos",
+			Namespace: ns,
+		},
+		Spec: v1alpha1.TimeChaosSpec{
+			Selector: v1alpha1.SelectorSpec{
+				Namespaces:     []string{ns},
+				LabelSelectors: map[string]string{"app": "timer"},
+			},
+			Mode:       v1alpha1.OnePodMode,
+			Duration:   pointer.StringPtr("9m"),
+			TimeOffset: "-1h",
+			Scheduler: &v1alpha1.SchedulerSpec{
+				Cron: "@every 10m",
+			},
+		},
+	}
+	err = cli.Create(ctx, timeChaos)
+	framework.ExpectNoError(err, "create time chaos error")
+
+	By("waiting for assertion")
+	err = wait.PollImmediate(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		podTime, err := getPodTimeNS(c, port)
+		framework.ExpectNoError(err, "failed to get pod time")
+		if podTime.Before(*initTime) {
+			return true, nil
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "time chaos doesn't work as expected")
+
+	By("delete chaos CRD objects")
+	err = cli.Delete(ctx, timeChaos)
+	framework.ExpectNoError(err, "failed to delete time chaos")
+	time.Sleep(10 * time.Second)
+
+	By("waiting for assertion recovering")
+	err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (done bool, err error) {
+		podTime, err := getPodTimeNS(c, port)
+		framework.ExpectNoError(err, "failed to get pod time")
+		// since there is no timechaos now, current pod time should not be earlier
+		// than the init time
+		if podTime.Before(*initTime) {
+			return true, nil
+		}
+		return false, nil
+	})
+	framework.ExpectError(err, "wait no timechaos error")
+	framework.ExpectEqual(err.Error(), wait.ErrWaitTimeout.Error())
+	By("success to perform time chaos")
+
+}
+
+func TestcaseTimeSkewPauseThenUnpause(
+	ns string,
+	cli client.Client,
+	c http.Client,
+	port uint16,
+) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	By("wait e2e helper ready")
+	err := util.WaitE2EHelperReady(c, port)
+	framework.ExpectNoError(err, "wait e2e helper ready error")
+
+	initTime, err := getPodTimeNS(c, port)
+	framework.ExpectNoError(err, "failed to get pod time")
+
+	By("create chaos CRD objects")
+	timeChaos := &v1alpha1.TimeChaos{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "timer-time-chaos",
+			Namespace: ns,
+		},
+		Spec: v1alpha1.TimeChaosSpec{
+			Selector: v1alpha1.SelectorSpec{
+				Namespaces:     []string{ns},
+				LabelSelectors: map[string]string{"app": "timer"},
+			},
+			Mode:       v1alpha1.OnePodMode,
+			Duration:   pointer.StringPtr("9m"),
+			TimeOffset: "-1h",
+			Scheduler: &v1alpha1.SchedulerSpec{
+				Cron: "@every 10m",
+			},
+		},
+	}
+	err = cli.Create(ctx, timeChaos)
+	framework.ExpectNoError(err, "create time chaos error")
+
+	By("waiting for assertion")
+	err = wait.PollImmediate(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		podTime, err := getPodTimeNS(c, port)
+		framework.ExpectNoError(err, "failed to get pod time")
+		if podTime.Before(*initTime) {
+			return true, nil
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "time chaos doesn't work as expected")
+
+	chaosKey := types.NamespacedName{
+		Namespace: ns,
+		Name:      "timer-time-chaos",
+	}
+
+	By("pause time skew chaos experiment")
+	// pause experiment
+	err = util.PauseChaos(ctx, cli, timeChaos)
+	framework.ExpectNoError(err, "pause chaos error")
+
+	By("assert pause is effective")
+	err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		chaos := &v1alpha1.TimeChaos{}
+		err = cli.Get(ctx, chaosKey, chaos)
+		framework.ExpectNoError(err, "get time chaos error")
+		if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhasePaused {
+			return true, nil
+		}
+		return false, err
+	})
+	framework.ExpectNoError(err, "check paused chaos failed")
+
+	// wait for 1 minutes and check timer
+	framework.ExpectNoError(err, "get timer pod error")
+	err = wait.Poll(5*time.Second, 1*time.Minute, func() (done bool, err error) {
+		podTime, err := getPodTimeNS(c, port)
+		framework.ExpectNoError(err, "failed to get pod time")
+		if podTime.Before(*initTime) {
+			return true, nil
+		}
+		return false, nil
+	})
+	framework.ExpectError(err, "wait time chaos paused error")
+	framework.ExpectEqual(err.Error(), wait.ErrWaitTimeout.Error())
+
+	By("resume time skew chaos experiment")
+	err = util.UnPauseChaos(ctx, cli, timeChaos)
+	framework.ExpectNoError(err, "resume chaos error")
+
+	By("assert chaos experiment resumed")
+	err = wait.Poll(5*time.Second, 5*time.Minute, func() (done bool, err error) {
+		chaos := &v1alpha1.TimeChaos{}
+		err = cli.Get(ctx, chaosKey, chaos)
+		framework.ExpectNoError(err, "get time chaos error")
+		if chaos.Status.Experiment.Phase == v1alpha1.ExperimentPhaseRunning {
+			return true, nil
+		}
+		return false, err
+	})
+	framework.ExpectNoError(err, "check resumed chaos failed")
+
+	// timechaos is running again, we want to check pod
+	// whether time is earlier than init time,
+	err = wait.PollImmediate(5*time.Second, 1*time.Minute, func() (done bool, err error) {
+		podTime, err := getPodTimeNS(c, port)
+		framework.ExpectNoError(err, "failed to get pod time")
+		if podTime.Before(*initTime) {
+			return true, nil
+		}
+		return false, nil
+	})
+	framework.ExpectNoError(err, "time chaos failed")
+
+	By("delete chaos CRD objects")
+	cli.Delete(ctx, timeChaos)
+}
